+++ src/upathlib/gcp.py
@@ -7,12 +7,11 @@ from __future__ import annotations
 import contextlib
 import logging
 import os
-# import socket
-# import urllib3
+import time
 from io import BufferedReader, UnsupportedOperation, BytesIO
 from typing import Union
 
-# import requests
+import opnieuw
 from google import resumable_media
 from google.oauth2 import service_account
 from google.cloud import storage
@@ -23,7 +22,6 @@ from overrides import overrides
 from ._upath import FileInfo, Upath, LockAcquisitionTimeoutError
 from ._blob import BlobUpath
 from ._local import LocalUpath
-from ._util import Backoff
 
 
 logger = logging.getLogger(__name__)
@@ -62,6 +60,9 @@ class GcpBlobUpath(BlobUpath):
 
         If this code is used on a GCP machine, already in the context of an account/project, then
         `project_id` and `credentials` are not needed (but not harmful either).
+
+        TODO: check out github repo `googleapis/python-cloud-core` to see whether there is
+        a way or a need to infer these values. In particular, see class `google.cloud.client.ClientWithProject`.
         '''
         if bucket_name is None:
             assert len(paths) == 1
@@ -189,41 +190,17 @@ class GcpBlobUpath(BlobUpath):
         except NotFound:
             return None
 
-    # def _blob_retry(self, func_name, *args, max_tries=3, **kwargs):
-    #     # `func_name` is the name of a blob method.
-    #     sleeper = Backoff(0.1)
-    #     while True:
-    #         try:
-    #             return getattr(self.blob(), func_name)(*args, **kwargs)
-    #         except (requests.exceptions.ReadTimeout,
-    #                 urllib3.exceptions.ReadTimeoutError,
-    #                 socket.timeout,
-    #                 urllib3.exceptions.SSLError,
-    #                 requests.exceptions.SSLError,
-    #                 requests.exceptions.ConnectionError,
-    #                 ) as e:
-    #             if sleeper.retries >= max_tries:
-    #                 raise
-    #             logger.info("retrying %r on error: %r", func_name, e)
-    #             sleeper.sleep()
-    #             # My hypothesis is that sometimes timeout happens
-    #             # because the cached `client`, `bucket`, `blob`
-    #             # have become stale, hence should be recreated.
-    #             self._client = None
-    #             self._bucket = None
-    #             self._blob = None
-
-    def _blob_rate_limit(self, func, *args, max_retries=5, **kwargs):
-        # `func_name` is the name of a create/update/delete function.
-        sleeper = Backoff(0.2)
-        while True:
-            try:
-                return func(*args, **kwargs)
-            except TooManyRequests as e:
-                if sleeper.retries >= max_retries:
-                    raise
-                logger.info("retrying %r on error: %r", func, e)
-                sleeper.sleep()
+    @opnieuw.retry(
+        retry_on_exceptions=(TooManyRequests, ),
+        max_calls_total=10,
+        retry_window_after_first_call_in_seconds=60,
+    )
+    def _blob_rate_limit(self, func, *args, **kwargs):
+        # `func_name` is a create/update/delete function.
+        # Google imposes rate limiting on such requests.
+        # According to Google doc, https://cloud.google.com/storage/quotas,
+        #   There is a write limit to the same object name. This limit is once per second.
+        return func(*args, **kwargs)
 
     @overrides
     def _copy_file(self, target: GcpBlobUpath, *, overwrite=False) -> None:
@@ -342,23 +319,30 @@ class GcpBlobUpath(BlobUpath):
             yield self / p[k:].rstrip('/')  # "subdirectories"
 
     def _acquire_lease(self, *, timeout: int = None):
+        # Note: `timeout = None` does not mean infinite wait.
+        # It means a default wait time. If user wants longer wait,
+        # just pass in a large number. Because user often associate
+        # `timeout = None` with infinite wait, the default wait
+        # is a long period.
         if self._path == '/':
             raise UnsupportedOperation("can not write to root as a blob", self)
         if timeout is None:
-            timeout = 600  # seconds
-        sleeper = Backoff(0.1)
-        while True:
-            try:
-                # b.upload_from_string(b'0', if_generation_match=0)
-                # b.cache_control = 'no-store'
-                # b.patch()
-                self._blob_rate_limit(self._write_bytes, b'0')
-                self._generation = self.blob().generation
-                return
-            except (PreconditionFailed, FileExistsError):
-                if (t := sleeper.time_elapsed) >= timeout:
-                    raise LockAcquisitionTimeoutError(self, t)
-                sleeper.sleep()
+            timeout = 3600  # seconds
+
+        @opnieuw.retry(
+            retry_on_exceptions=(PreconditionFailed, FileExistsError),
+            max_calls_total=10,
+            retry_window_after_first_call_in_seconds=timeout,
+        )
+        def _acquire_():
+            self._blob_rate_limit(self._write_bytes, b'0')
+            self._generation = self.blob().generation
+
+        t0 = time.perf_counter()
+        try:
+            _acquire_()
+        except Exception as e:
+            raise LockAcquisitionTimeoutError(self, time.perf_counter() - t0) from e
 
     @contextlib.contextmanager
     @overrides
@@ -386,7 +370,6 @@ class GcpBlobUpath(BlobUpath):
                         self.blob().delete,
                         client=self.client,
                         if_generation_match=self._generation,
-                        max_retries=10,
                     )
                 except Exception as e:
                     logger.error(e)